requirements.txt
Flask==2.1.1
psycopg2-binary==2.8.6
joblib==0.17.0
pandas==1.4.2
sklearn==0.0
boto3==1.24.34
requests==2.28.1
jsonify==0.5
sagemaker==2.100.0